Logging *should* work now
There is a bug with the level logic which makes `JSONSerialsizer.Instance.SceneNumber` report two numbers (the second always being 7) each time a new level is loaded. I believe this is causing the logger to always write to 7.csv instead of the proper trial nums. (I just realized that it should be based off of trial num, not level num... But still this is a bug that has to be fixed for proper stats)